remove test_jit_py3.py
1. move some to module
2. created test_types for types and annotation
3. created test_misc for the rest
